by.bsuir.radionov/task7/src/Task7.java
import java.util.Arrays;

public class Task7 {
    public static void main(String[] args) {
        double[] mas = {4, 3, 2, 1, 52, 21, 14, 3152, 163, 112, 511, 12, 444, 124, 1, 4, 124, 567};
        bubbleSort(mas);
        System.out.println(Arrays.toString(mas));
    }

    public static void bubbleSort(double[] arr) {
        int n = arr.length;
        boolean swapped;

        do {
            swapped = false;
            for (int i = 0; i < n - 1; i++) {
                if (arr[i] > arr[i + 1]) {
                    swap(arr, i, i + 1);
                    swapped = true;
                }
            }
        } while (swapped);
    }

    private static void swap(double[] array, int j, int i) {
        double temp = array[i];
        array[i] = array[j];
        array[j] = temp;
    }
}
